camel-main: move Option from MainSupport to MainCommandLineSupport
